Move transaction implementations to nested package
