+++ java/compaction/compaction-core/src/main/java/sleeper/compaction/core/job/CompactionJobCommitterOrSendToLambda.java
@@ -24,7 +24,7 @@ import sleeper.core.properties.table.TablePropertiesProvider;
 import sleeper.core.statestore.ReplaceFileReferencesRequest;
 import sleeper.core.statestore.StateStoreProvider;
 import sleeper.core.statestore.commit.StateStoreCommitRequest;
-import sleeper.core.statestore.transactionlog.transaction.ReplaceFileReferencesTransaction;
+import sleeper.core.statestore.transactionlog.transaction.impl.ReplaceFileReferencesTransaction;
 import sleeper.core.table.TableStatus;
 import sleeper.core.tracker.compaction.job.CompactionJobTracker;
 import sleeper.core.tracker.compaction.job.update.CompactionJobFinishedEvent;
