+++ java/compaction/compaction-core/src/main/java/sleeper/compaction/core/job/commit/CompactionCommitBatcher.java
@@ -20,7 +20,7 @@ import org.slf4j.LoggerFactory;
 
 import sleeper.core.statestore.commit.StateStoreCommitRequest;
 import sleeper.core.statestore.commit.StateStoreCommitRequestSender;
-import sleeper.core.statestore.transactionlog.transaction.ReplaceFileReferencesTransaction;
+import sleeper.core.statestore.transactionlog.transaction.impl.ReplaceFileReferencesTransaction;
 
 import java.util.List;
 import java.util.Map;
