+++ java/core/src/main/java/sleeper/core/statestore/transactionlog/transaction/TransactionType.java
@@ -15,6 +15,15 @@
  */
 package sleeper.core.statestore.transactionlog.transaction;
 
+import sleeper.core.statestore.transactionlog.transaction.impl.AddFilesTransaction;
+import sleeper.core.statestore.transactionlog.transaction.impl.AssignJobIdsTransaction;
+import sleeper.core.statestore.transactionlog.transaction.impl.ClearFilesTransaction;
+import sleeper.core.statestore.transactionlog.transaction.impl.DeleteFilesTransaction;
+import sleeper.core.statestore.transactionlog.transaction.impl.InitialisePartitionsTransaction;
+import sleeper.core.statestore.transactionlog.transaction.impl.ReplaceFileReferencesTransaction;
+import sleeper.core.statestore.transactionlog.transaction.impl.SplitFileReferencesTransaction;
+import sleeper.core.statestore.transactionlog.transaction.impl.SplitPartitionTransaction;
+
 /**
  * A marker for the type of a transaction. Used when storing a log entry for a transaction. Stabilises mapping between
  * a serialised transaction type and the class of the transaction, in case the class is renamed or moved.
