+++ java/core/src/main/java/sleeper/core/statestore/transactionlog/transaction/impl/SplitFileReferencesTransaction.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.core.statestore.transactionlog.transaction;
+package sleeper.core.statestore.transactionlog.transaction.impl;
 
 import sleeper.core.statestore.FileReference;
 import sleeper.core.statestore.SplitFileReferenceRequest;
@@ -24,6 +24,7 @@ import sleeper.core.statestore.exception.FileReferenceAssignedToJobException;
 import sleeper.core.statestore.exception.FileReferenceNotFoundException;
 import sleeper.core.statestore.transactionlog.state.StateStoreFile;
 import sleeper.core.statestore.transactionlog.state.StateStoreFiles;
+import sleeper.core.statestore.transactionlog.transaction.FileReferenceTransaction;
 
 import java.time.Instant;
 import java.util.List;
