+++ java/core/src/test/java/sleeper/core/statestore/transactionlog/transactions/TransactionSerDeTest.java
@@ -30,19 +30,19 @@ import sleeper.core.statestore.AllReferencesToAFile;
 import sleeper.core.statestore.FileReference;
 import sleeper.core.statestore.FileReferenceFactory;
 import sleeper.core.statestore.ReplaceFileReferencesRequest;
-import sleeper.core.statestore.transactionlog.transaction.AddFilesTransaction;
-import sleeper.core.statestore.transactionlog.transaction.AssignJobIdsTransaction;
-import sleeper.core.statestore.transactionlog.transaction.ClearFilesTransaction;
-import sleeper.core.statestore.transactionlog.transaction.DeleteFilesTransaction;
 import sleeper.core.statestore.transactionlog.transaction.FileReferenceTransaction;
-import sleeper.core.statestore.transactionlog.transaction.InitialisePartitionsTransaction;
 import sleeper.core.statestore.transactionlog.transaction.PartitionTransaction;
-import sleeper.core.statestore.transactionlog.transaction.ReplaceFileReferencesTransaction;
-import sleeper.core.statestore.transactionlog.transaction.SplitFileReferencesTransaction;
-import sleeper.core.statestore.transactionlog.transaction.SplitPartitionTransaction;
 import sleeper.core.statestore.transactionlog.transaction.StateStoreTransaction;
 import sleeper.core.statestore.transactionlog.transaction.TransactionSerDe;
 import sleeper.core.statestore.transactionlog.transaction.TransactionType;
+import sleeper.core.statestore.transactionlog.transaction.impl.AddFilesTransaction;
+import sleeper.core.statestore.transactionlog.transaction.impl.AssignJobIdsTransaction;
+import sleeper.core.statestore.transactionlog.transaction.impl.ClearFilesTransaction;
+import sleeper.core.statestore.transactionlog.transaction.impl.DeleteFilesTransaction;
+import sleeper.core.statestore.transactionlog.transaction.impl.InitialisePartitionsTransaction;
+import sleeper.core.statestore.transactionlog.transaction.impl.ReplaceFileReferencesTransaction;
+import sleeper.core.statestore.transactionlog.transaction.impl.SplitFileReferencesTransaction;
+import sleeper.core.statestore.transactionlog.transaction.impl.SplitPartitionTransaction;
 import sleeper.core.util.NumberFormatUtils;
 
 import java.time.Instant;
