+++ java/splitter/splitter-core/src/main/java/sleeper/splitter/core/split/SplitPartition.java
@@ -24,7 +24,7 @@ import sleeper.core.schema.Field;
 import sleeper.core.schema.Schema;
 import sleeper.core.statestore.StateStore;
 import sleeper.core.statestore.commit.StateStoreCommitRequest;
-import sleeper.core.statestore.transactionlog.transaction.SplitPartitionTransaction;
+import sleeper.core.statestore.transactionlog.transaction.impl.SplitPartitionTransaction;
 import sleeper.splitter.core.split.FindPartitionSplitPoint.SketchesLoader;
 
 import java.util.List;
