+++ java/statestore-lambda/src/test/java/sleeper/statestore/lambda/committer/CompactionCommitBatcherLambdaIT.java
@@ -33,7 +33,7 @@ import sleeper.core.statestore.FileReferenceFactory;
 import sleeper.core.statestore.ReplaceFileReferencesRequest;
 import sleeper.core.statestore.commit.StateStoreCommitRequest;
 import sleeper.core.statestore.commit.StateStoreCommitRequestSerDe;
-import sleeper.core.statestore.transactionlog.transaction.ReplaceFileReferencesTransaction;
+import sleeper.core.statestore.transactionlog.transaction.impl.ReplaceFileReferencesTransaction;
 import sleeper.localstack.test.LocalStackTestBase;
 
 import java.util.List;
