+++ java/statestore/src/main/java/sleeper/statestore/s3/S3FileReferenceStore.java
@@ -32,12 +32,12 @@ import sleeper.core.statestore.exception.ReplaceRequestsFailedException;
 import sleeper.core.statestore.exception.SplitRequestsFailedException;
 import sleeper.core.statestore.transactionlog.state.StateStoreFile;
 import sleeper.core.statestore.transactionlog.state.StateStoreFiles;
-import sleeper.core.statestore.transactionlog.transaction.AddFilesTransaction;
-import sleeper.core.statestore.transactionlog.transaction.AssignJobIdsTransaction;
-import sleeper.core.statestore.transactionlog.transaction.DeleteFilesTransaction;
 import sleeper.core.statestore.transactionlog.transaction.FileReferenceTransaction;
-import sleeper.core.statestore.transactionlog.transaction.ReplaceFileReferencesTransaction;
-import sleeper.core.statestore.transactionlog.transaction.SplitFileReferencesTransaction;
+import sleeper.core.statestore.transactionlog.transaction.impl.AddFilesTransaction;
+import sleeper.core.statestore.transactionlog.transaction.impl.AssignJobIdsTransaction;
+import sleeper.core.statestore.transactionlog.transaction.impl.DeleteFilesTransaction;
+import sleeper.core.statestore.transactionlog.transaction.impl.ReplaceFileReferencesTransaction;
+import sleeper.core.statestore.transactionlog.transaction.impl.SplitFileReferencesTransaction;
 import sleeper.statestore.StateStoreArrowFileStore;
 
 import java.io.IOException;
